// src/generated-sources/openapi/model/accountResource.ts
/**
 * Up API
 * The Up API gives you programmatic access to your balances and transaction data. You can request past transactions or set up webhooks to receive real-time events when new transactions hit your account. It’s new, it’s exciting and it’s just the beginning. 
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from '../api';
import { AccountResourceAttributes } from './accountResourceAttributes';
import { AccountResourceLinks } from './accountResourceLinks';
import { AccountResourceRelationships } from './accountResourceRelationships';

/**
* Provides information about an Up bank account. 
*/
export class AccountResource {
    /**
    * The unique identifier for this account. 
    */
    'id': string;
    /**
    * The type of this resource: `accounts`
    */
    'type': string;
    'attributes': AccountResourceAttributes;
    'links'?: AccountResourceLinks;
    'relationships': AccountResourceRelationships;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "string"
        },
        {
            "name": "type",
            "baseName": "type",
            "type": "string"
        },
        {
            "name": "attributes",
            "baseName": "attributes",
            "type": "AccountResourceAttributes"
        },
        {
            "name": "links",
            "baseName": "links",
            "type": "AccountResourceLinks"
        },
        {
            "name": "relationships",
            "baseName": "relationships",
            "type": "AccountResourceRelationships"
        }    ];

    static getAttributeTypeMap() {
        return AccountResource.attributeTypeMap;
    }
}