// src/generated-sources/openapi/model/accountResourceRelationships.ts
/**
 * Up API
 * The Up API gives you programmatic access to your balances and transaction data. You can request past transactions or set up webhooks to receive real-time events when new transactions hit your account. It’s new, it’s exciting and it’s just the beginning. 
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from '../api';
import { AccountResourceRelationshipsTransactions } from './accountResourceRelationshipsTransactions';

export class AccountResourceRelationships {
    'transactions': AccountResourceRelationshipsTransactions;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "transactions",
            "baseName": "transactions",
            "type": "AccountResourceRelationshipsTransactions"
        }    ];

    static getAttributeTypeMap() {
        return AccountResourceRelationships.attributeTypeMap;
    }
}