// src/generated-sources/openapi/model/cashbackObject.ts
/**
 * Up API
 * The Up API gives you programmatic access to your balances and transaction data. You can request past transactions or set up webhooks to receive real-time events when new transactions hit your account. It’s new, it’s exciting and it’s just the beginning. 
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from '../api';
import { MoneyObject } from './moneyObject';

/**
* Provides information about an instant reimbursement in the form of cashback. 
*/
export class CashbackObject {
    /**
    * A brief description of why this cashback was paid. 
    */
    'description': string;
    /**
    * The total amount of cashback paid, represented as a positive value. 
    */
    'amount': MoneyObject;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "description",
            "baseName": "description",
            "type": "string"
        },
        {
            "name": "amount",
            "baseName": "amount",
            "type": "MoneyObject"
        }    ];

    static getAttributeTypeMap() {
        return CashbackObject.attributeTypeMap;
    }
}