// src/generated-sources/openapi/model/errorObject.ts
/**
 * Up API
 * The Up API gives you programmatic access to your balances and transaction data. You can request past transactions or set up webhooks to receive real-time events when new transactions hit your account. It’s new, it’s exciting and it’s just the beginning. 
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from '../api';
import { ErrorObjectSource } from './errorObjectSource';

/**
* Provides information about an error processing a request. 
*/
export class ErrorObject {
    /**
    * The HTTP status code associated with this error. This can also be obtained from the response headers. The status indicates the broad type of error according to HTTP semantics. 
    */
    'status': string;
    /**
    * A short description of this error. This should be stable across multiple occurrences of this type of error and typically expands on the reason for the status code. 
    */
    'title': string;
    /**
    * A detailed description of this error. This should be considered unique to individual occurrences of an error and subject to change. It is useful for debugging purposes. 
    */
    'detail': string;
    'source'?: ErrorObjectSource;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "status",
            "baseName": "status",
            "type": "string"
        },
        {
            "name": "title",
            "baseName": "title",
            "type": "string"
        },
        {
            "name": "detail",
            "baseName": "detail",
            "type": "string"
        },
        {
            "name": "source",
            "baseName": "source",
            "type": "ErrorObjectSource"
        }    ];

    static getAttributeTypeMap() {
        return ErrorObject.attributeTypeMap;
    }
}